requirements.txt
openai>=1.0.0
langchain>=0.1.0
chromadb>=0.4.0
fastapi>=0.104.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
gitpython>=3.1.0
uvicorn>=0.24.0
websockets>=12.0.0

# Deployment
pyinstaller>=5.0.0
